Feat: adição de JSDocs sobre os serviços remotos e pequenos ajustes
