+++ src/services/permissions.js
@@ -1,6 +1,11 @@
 import api from './base.js';
 
 export default {
+  /**
+   * GET /v1/permissions
+   *
+   * @param {object} params
+   */
   async index (params = {}) {
     const { data } = await api.get('/permissions', { params })
     return data
