+++ src/services/sector.js
@@ -1,23 +1,53 @@
 import api from './base';
 
 export default {
+  /**
+   * GET /v1/sectors
+   *
+   * @param {object} params
+   * @param {string} params.filter_id
+   * @param {string} params.find_name
+   * @param {string} params.sort_by
+   * @param {string} params.sort_order
+   */
   async index (params = {}) {
     const { data } = await api.get('/sectors', { params })
     return data
   },
 
-  async destroy (id) {
-    const { data } = await api.delete(`/sectors/${id}`)
+  /**
+   * POST /v1/sectors
+   *
+   * @param {object} body
+   * @param {string} body.name
+   * @param {number} body.project_id
+   */
+  async store (body) {
+    const { data } = await api.post('/sectors', body)
     return data
   },
 
+  /**
+   * PUT /v1/sectors/{id}
+   *
+   * @param {number} id
+   * @param {object} body
+   * @param {string} body.name
+   * @param {number} body.project_id
+   * @param {number} body.user_id
+   */
   async update (id, body) {
     const { data } = await api.put(`/sectors/${id}`, body)
     return data
   },
 
-  async store (body) {
-    const { data } = await api.post('/sectors', body)
+  /**
+   * DELETE /v1/sectors/{id}
+   *
+   * @param {number} id
+   */
+  async destroy (id) {
+    const { data } = await api.delete(`/sectors/${id}`)
     return data
   }
 }
